Relocate mixins
