+++ src/mixins/index.ts
@@ -1,0 +1,2 @@
+export { default as gatewayMixin } from './gatewayMixin';
+export { default as serviceMixin } from './serviceMixin';
